Migrate code_gen_python to code gen library
